Convert text files newlines to CR-LF, remove unused wolf3d APOGEE objects copies
Split to submodules note: This was the commit message
after git conversion and before split to submodules:

"Convert text files newlines to CR-LF, add updated wolf3d.diff, remove unused wolf3d APOGEE objects copies"

Mercurial to git conversion note: Initially, a single Mercurial
repository was used, with a directory named wolfsrc-lastvers.
It was renamed wolf3d in this commit, along with the aforementioned
patch file. As a part of a later conversion to git, it was decided
to use the name of wolf3d from the very beginning for most.