More breaking file-management related changes:
New locations are used for files now. The following locations are
the defaults, and can be overridden using the -datadir and -cfgdir
command line arguments. Assuming the defaults:
- On Windows, anything is written in %APPDATA%\reflection-keen.
- On Linux, new cfg files like refkdreams.cfg are written
in $XDG_CONFIG_HOME/reflection-keen or ~/.config-reflection-keen,
while other files (including saved games and config.*) are written
in $XDG_DATA_HOME/reflection-keen or ~/.local/share/reflection-keen.
The XDG paths have a higher priority.

For each game version, a separate subdirectory should be created within
the data subdirectory, e.g., kdreamse113 for Keen Dreams v1.13.
Inside it, a "writable" subdirectory should be created for
files a specific game may read and write, like saved games.
An additional "readonly" subdirectory may be present for game data,
including resources extracted from an original EXE (like egahead.kdr).
Such a file is not read from the "readonly" directory if it's already
found in the corresponding game installation's directory. However, if
the wrong file is found in "readonly" for a given game version, it
may be REMOVED (and possibly replaced for extracted sources).