It is now possible (with some limitations) to select a game directory.
For the sake of simplicity (mostly in terms of UI and cfg), only
directory names exclusively consisting of ASCII chars are shown,
and may be selected. Even the directory names in the newly generated
cfg files cannot be manually edited to have non-ASCII chars.
However, when such a directory is located from the launcher,
the first thing to choose is a "root path". Possible root paths
include all possible drives on Windows (identified by drive letters),
the path of / on Linux, the home dir on both platforms,
an installation dir of Steam and a default
installation dir of GOG.com games.