Add "key repeat"-like behaviors for game controller. Mouse is missing.
The case of using the mouse to click on on-screen buttons is still
not covered. It may be better to revert a lot of the changes to
be_st_sdl.c and be_st_sdl_graphics.c, and just track the last
emulated key press (by DOS scan code).
The launcher itself should still track gamepad buttons as
done now, though.