More key-repeat changes, including some code reverts.
Emulated key repeat with DOS scancode is now handled independently of
the actual input device in use. Thus, not only keyboard and controller
key/button presses are covered, but also mouse button presses with
on-screen controls. In particular, any OS/hardware level keyboard
repeat that may be observed is ignored. These are emulated in software
using specific repeat delay and rate.

The latter is not done with keyboard input in the launcher, though.

As a side-note, binary threshold for game controller's analog features
has been changed.