User input related updates in the launcher
- Added a basic implementation of search-as-you-type, where it is
possible to begin typing the text forming a menu item label in order
to locate it. Done with directory selection in mind and usable with
a keyboard only (not just e.g., a game controller). For the sake of
simplicity and consistency with the game itself, US layout is further
assumed (only SDL scancodes are used, along with shift key flags).
- Key repeat is now emulated in the launcher in a similar manner
to what's done with the controller, as well as both
the keyboard and the controller in-game.
- Implemented page up/down, again done with directory selection in mind.
In the case of a game controller, shoulder buttons can be used.
- Better vertical scrolling with keyboard or game controller, skipping
static labels in the way while still doing a small scroll
if no selectable menu item is found.
- Only the primary mouse button can be used for selecting menu items in
this revision (a mistaken click on mouse wheel can lead to confusion).
- Improved vertical scrolling with the mouse, so it's somewhat more
natural after dragging (i.e., clicking-and-moving the mouse cursor).
Timing is also taken into account now, so the amount of calls to
BE_Launcher_RefreshVerticalScrolling isn't what defines the rate
of scrolling alone. Unfortunately this means that values returned
by SDL_GetTicks() are exposed to be_launcher.c, but this makes things
somewhat less messy, and there's still no direct reference to SDL.