EGA and CGA versions of Keen Dreams can be started from a single binary.
This has been postponed for a while, compared to the other versions'
mergers, but it should be done by now. This may be the trickiest one
in a few ways.
- Obviously, there are the different CGA and EGA implementations of
functions, mostly in ID_VW* and ID_RF*, that should be handled.
This includes the usage of function pointers, as well as variables
for integer values. In case it is clear which version of a function
or variable should be accessed, though, it may be done directly.
- CGA version 1.05 is also closer to the original, combined CGA+EGA
release from id Software, with things like the scrolling titlescreen,
as well as expecting just the title argument for CA_CacheMarks.
- Back to actual CGA vs. EGA differences, a case possibly being a bit
more tricky than usual is probably the need to split the spritetype
struct type into separate types, based on the different MAXSHIFTS
values. New macros are now in use for accessing the struct's members
in a video mode-agnostic way.