Call PrepareGamePlayControllerMapping in GameLoop instead of PlayLoop.
This gives the user a bit more time to use the controller for in-game
actions (e.g., moving the character around) as the game begins, at least
in Keen Dreams. This still does not change the fact that the controllers
are not exact replicas of keyboards and have certain different
behaviors (like having the various interal controller mappings).
In addition, PrepareGamePlayControllerMapping may still be called
from PlayLoop, via CheckKeys, after a call to US_ControlPanel
is complete (Keen Dreams, Catacomb 3-D).