Added a manualgamevermode setting inaccessible from the launcher.
This can be used if one wants to partially bypass the checksum
and filesize verifications done for game installations. It is suited
to reproduce the case where the game EXE (which is still verified)
remains the same, but game data files may differ. As a bonus, data
originally embedded in the EXE (like EGAHEAD.KDR) may also be replaced.

Note that this disables the automatic search of known game
installations. Reason is, if such a game installation is
found and added, there's no way to add another directory
with the same version of the game EXE.

This is considered an unofficial feature, since it is expected to be
quite obscure anyway, and unexpected behaviors may arrive if used
in certain ways. For instance, usually, wrong game data files
are removed from e.g., ~/.local/share/reflection-keen/*, but
such files are acceptable with manualgamevermode=true
(except for the game EXE).