Alright, let's enable the modern controller scheme by default.
While it does sound nice to behave more like the original DOS versions
as a default, it does not make sense in practice. It isn't that
common, these days, to use a 2-buttons joystick, or even a
4-buttons gamepad (with a single d-pad, 4 buttons and no more).
Furthermore, not anything can be done with these joysticks and gamepads
in the same DOS versions. The keyboard is still required for a lot
of functionalities. Here, the modern game controller scheme aims
to make it possible to use a game controller for anything,
although a few bits may still be missing.