Start (most) of conversion of g_sdlVidMem.egaGfx to linear pattern.
This is considered mostly beneficial for drawing scaled walls and
shapes in the Catacombs, since EGA Write mode 2 was originally
used. This basically means the input in use was a color in
the range 0-15, which could fill up to 8 pixels.

Furthermore, BEL_ST_UpdateHostDisplay seemed to have a clear
part in spending CPU time, given the need to to convert the above
egaGfx data from planar format to linear indexed color format
(and then to 32-bit "true color" format).

While storing separate planes may appear to be more beneficial
for functions like BE_ST_EGAUpdateGFXBufferInPlane, improvements
were observed in Keen Dreams, too; Mostly in the menu, but also
during gameplay, probably since less work has to be done
in BEL_ST_UpdateHostDisplay.

Known limitation: The EGA line width property should *not* be shorter
than the actual displayed width, otherwise things may not be drawn as
as expected (although it looks like this never happens in practice).