Now using lookup tables for the new EGA data arrangement.
This also includes making the "MemCopy" family of functions look
a bit less messy. Unfortunately, we're talking about 3 new tables,
where each of them has the size of 2KiB, but that's not as much as
doubling the size of g_sdlVidMem.egaGfx from 256KiB to 512KiB.

The usage of the lookup tables may prove to be somewhat better
with 64-bit architectures, but it doesn't mean
it's bad to use them otherwise.