Time to add initial touch input support! This includes a few adjustments to the C codebase for Android.
Things to consider:
- There's no proper support for multi-touch, if any of the on-screen
controls previously added for game controllers support
(e.g., an on-screen keyboard) is in use.
- Usage of BE_MENUITEM_TYPE_SELECTION menu items
with touch input can be improved.
- Mouse button/Finger press can select a menu item in any of the
launcher's menus even if done outside the menus' boundaries.
- Keen Dreams control panel: Keyboard emulation is currently
used for touch input. At the least, this works
even if the /NOMOUSE parameter is passed.
- In-game on-screen d-pad is technically split into 8 areas to
take care of diagonal movement. Unfortunately, at the moment
this has the side-effect of using just a part of the d-pad
when diagonal movement is irrelevant.
- While the game controller related UI controls were resized,
the d-pad's appearance is still a little bit different, compared
to the way it appears on-screen for in-game character movement.
- Touch controls' textures maintenance (e.g., creation) can be improved.
- Currently touch input can either be permanently disabled in-game, or
permanently enabled. In the latter case, a mouse can be used like a finger.

Misc Android-related notes:
- Touch input is enabled on Android by default,
and disabled on other platforms (by default).
- The default sample rate is now 48000Hz, because 49716Hz may lead to
unexpected behaviors (e.g., a hang) on Android.
- The SDL2 scancode of SDL_SCANCODE_AC_BACK (back button on Android)
is internally translated to the DOS scancode of the Escape key in-game.
- be_cross.c: BE_Cross_LogMessage writes to the logcat on Android.
- be_cross_fs.c: SDL_AndroidGetExternalStoragePath() is used as the location
of files internally read and written by the refkeen ports. The environment
variables EXTERNAL_STORAGE and EXTERNAL_SDCARD_STORAGE are used to get basic
so-called "root paths" for game installations selections in the launcher
(although only EXTERNAL_STORAGE has been tested to work on one device).
Note that these environment variables don't appear to reside in any official
documentation, and there may be no official way to get the correct paths
(at least before a certain Android API version). However, these seem to work
on various devices and are simple to use from C/C++.