Horizontally swap the order of the on-screen touch controls on the top half of the display, for the Catacombs.
This is done so, while one finger is used to move the character,
another one can be used to move/rotate fast or use a powersphere.
Low-priority controls (like debug keys) may appear above
the analog circle.

The locations of the controls in Keen Dreams were not changed,
since the need is not there, while the low-priority controls
could partially cover the scorebox if relocated to the left.