ANOTHER BREAKING CHANGE: Replace "autolock" with "mousegrab" setting.
This should be similar to the "Grab mouse in windowed mode" toggle
from Chocolate Doom, although there are a few differences.
To begin with, there's no change in behaviors for fullscreen windows.

In windowed mode, these are the possible setting values and
their meanings:
- auto: Ungrab the mouse if it doesn't have to be grabbed. For
instance, if the mouse is enabled, then it'll be grabbed in
Catacomb 3-D during gameplay, while it won't be in the
DemoLoop sequence (e.g., the high scores table).
- off: Permanently ungrab the mouse.
- commonly: Almost always grab the mouse. It'll be ungrabbed if
on-screen control (e.g., touch controls) are shown, or in case
the touchinput setting is set to "forced".

Known caveats:
- A recent SDL2 build should be used, as otherwise the mouse cursor may
appear to unexpectedly jump after entering the window from outside.
A similar case is where mouse grab is automatically toggled OFF,
say after leaving a scrolling text sequence in
the Keen Dreams control panel.