Add the "absmousemotion" setting (Keen Dreams specific).
When toggled on, this makes it possible to seamlessly move the mouse
cursor out of the window while in the Keen Dreams control panel.
This is similar to the way motion is handled with touch input.

In addition, g_sdlMouseButtonsStates and g_sdlEmuMouseButtonsState are
now reset only when on-screen UI is shown/hidden, and not always
whenever the mouse is grabbed/ungrabbed.