Add support for speexdsp on Android, and set speexdsp to the default resampling library, only since it's available from an Ubuntu 12.04 repository (even though SDL2 isn't).
This adaptation of speexdsp to Android has been used:
https://github.com/xyx119/speexdsp-ndk

src/android-lib/jni/speexdsp should be a copy of (or a symlink to)
the jni subdirectroy in the repository.

Again, Android.mk shall be edited, so $(BUILD_SHARED_LIBRARY) is
included.

Unfortunately, the folder name of speexdsp-1.2rc3 is currently
hardcoded in the Android.mk files.