Move common Android.mk code to a new Android_common.mk file (similar to Makefile.common), and apply minor fixes (resampling libraries' directories).
In addition, it looks like there's no need to explicitly load the resampling
library in use, at least if libmain.so is loaded. So a little piece of code
can be removed from RefkeenActivity.java, and it doesn't have to be edited
in order to use another resampling library (or disable resampling altogether).