Added support for the game data from the 2015 re-release of Keen Dreams (Steam, IndieGameStand).
Please note that despite all the modifications/additions done to
support this data, the support itself should still be considered
a bonus. In particular, this is *not* a direct derivative of
the 2015 port aiming to behave the same where possible, and
differences are expected (more than with the DOS versions).

- Joysticks / Game controllers are supported in the same way as with
data from the DOS versions.
- Mouse support is also the same, although it's possible to pass
the command-line argument of /SWMOUSE to temporarily disable
the "Absolute mouse motion" toggle. Note that the /NOASPECT and
/WINDOWED arguments described in GAMETEXT.KDR didn't seem to work
in the 2015 port itself (using an EXE from May 9 2016 or earlier),
so these are not in reflection-kdreams as of this commit.
- For now, while a more suitable function name may be desirable,
the function BE_Cross_free_mem_loaded_embedded_rsrc is also used for
loading external files from the 2015 data, because it basically works.
- Macros like GRHEADERLINKED are still defined, and the linkage of
data files like SOUNDHHD.KDR is faked, since it was
more convenient to do so.
- With the 2015 data, it's possible to dynamically switch between EGA
and CGA graphics. Note that the CGA graphics are actually "Fake CGA",
since they are internally stored as EGAGRAPH data, and
a different palette (fakecgacolors) is used.
- As a side-effect, variables like "F_FIRSTCOLOR" can be dynamically
changed, and id_ca.c has new functions like CA_ReloadGrChunks.
- Added a new "fakecgamode" variable, which is *always* set to false
with the DOS versions. In fact, there's often a check of this variable
*without* querying the game version. Again, the DOS versions
imply fakecgamode==false. Otherwise, this is basically
the EGA / (Fake) CGA toggle from the 2015 port.
- id_mm.c also has a couple of new helper functions, used by
CA_ReloadGrChunks. As for memory management, with the 2015 data,
MM_GetPtr calls "malloc", FREEBLOCK includes a call to "free",
and memory blocks aren't necessarily sorted by useptr.
- Support for digitized sound playback was added. The 2015 data
includes Huffman-compressed DPCM sample data with the frequency of
44100Hz and the sample format of int16_t. It can be said that
the DPCM data consists of second-order differences of samples.
- id_sd.c was modified for support of digitized sound playback, but
most changes were really done to be_st_sdl_audio_timer.c. Audio
resampling can still be used, although it's the rate of 44100Hz
which makes resampling unnecessary (rather than 49716Hz).
Additionally, it's still possible to have e.g., SD_WaitSoundDone
working as expected while the sound subsystem is disabled.
- be_cross_fs.c also has minor changes, regarding digitized sound
playback support. Basically, each game version has its own
"sample rate" definition. Usually it is unused and set to 0,
but the value of 44100 is assigned to the 2015 data.
- Of course, be_cross_fs.c has the changes required for supporting
the 2015 data in general (not just the digitized sounds rate field).
This includes an attempt to auto-detect and use a Steam installation.
- The CREDITS macro was replaced with a variable, which is set to 1
for the 2015 port. Two credits screen from KDREAMS.CMP are shown
in DemoLoop, *if* actual EGA graphics are used.
- As in the 2015 port, gamestate.keys is reset to 0 when
a new game is started. However, this is still *not* done
when any of the DOS versions' data is used.
- Added BE_ST_SetAppQuitCallback, so when an SDL_QUIT event arrives,
the kdreams.cfg file is updated, but only if the 2015 data is used.
- There's partial compatibility with saved games from the 2015 port.
The various states' offsets should be the same as in v1.13, although
the saved game files aren't compatible with DOS version 1.13. There's
also no guarantee that behaviors will be exactly the same across
the 2015 port and reflection-kdreams.
- There's further partial compatibility with the kdreams.cfg file.
One great gotcha is the way the keys' scancodes are stored.
The 2015 port uses SDL2 scancodes. For the sake of simplicity,
these are converted to DOS scancodes on file read,
and then converted back on file write.
- Unfortunately, conversions as done above can easily lead to different
contents, e.g., SDL_SCANCODE_KP_4 changing to SDL_SCANCODE_LEFT.
- The fullscreen and aspect ratio settings are still read/written
from/to reflection-kdreams.cfg, as usual, but they're *additionally*
written to the kdreams.cfg file on quit (2015 data only).

Other updates:

- BE_ST_InitAudio is now called after BE_ST_PrepareForGameStartup,
leading to a rename to BE_ST_PrepareForGameStartupWithoutAudio.
Reason is that an error message may be displayed, if there's a need.
- Misc. id_mm.c changes.
- Fixes for NULL pointer dereferences in ChangeState and ClipToWalls;
Reproduced when loading a saved game in which a vegatable is in the
middle of transforming from flower *back* to its original form.
Hopefully this is closer to the behaviors under DOS.