Add the 1992-11-12 release of SOD, currently under SODFOR13.PRJ.
The project file is essentially a copy of SODFOR10.PRJ, modified to
have the macro GAMEVER_RESTORATION_SOD_FOR14 defined (instead of FOR10),
and built with Borland C++ 3.1 (rather than 3.0 as used for SOD v1.0).
The SIGNON.OBJ object for v1.0 is still used, though.

To summarize, this is SODFOR14.PRJ, the 1993-01-12 v1.4 release,
with a few properties of SODFOR10.PRJ, the 1992-09-28 v1.0 release.