Add new GAMEVER_NOAH3D and GAMEVER_WOLFREV macros (currently unused).
The plan is to refrain from using the GAMEVER_EXEDEF* macros anywhere
other than VERSION.*, while GAMEVER_WOLFREV will be used to tell a part
changes between different revisions of the codebase. Modifications
specific to S3DNA, which are unrelated to the changes as defined
by GAMEVER_WOLFREV (and essentially belong to a separate "branch")
are to be given by GAMEVER_NOAH3D (currently in a somewhat similar
fashion to macros like SPEAR and GOODTIMES).