OK, following some feedback, let's name the various revisions.
To reduce any confusion, these are directly based on the names of
the projects. For each project, there's now a corresponding GV_WR
macro, like GV_WR_WL1AP10 (currently defined to 19920505L).

For the C sources only (hence excluding ASM), in addition to the
usual GAMEVER_EXEDEF_*** macros, there's now a separate GAMEVER_EXEDEF
macro. It is defined to an expression like WL1AP10, to be concatenated
to GV_WR in VERSION.H. This is done so GAMEVER_WOLFREV will be defined
to a macro like GV_WL_WL1AP10.

Since it might be impossible to do macro concatenations like
this in Turbo Assembler, there's a list of ifdefs in VERSION.EQU
for now. The original GAMEVER_EXEDEF_*** macros are still used here.

Note that GAMEVER_WOLFREV is still compared to the bare (long) integer
literal of 19921007L in ID_PM.C, for the AJR bugfix.