The beginning of be_cross_mem.c.
This has two purposes:
- Move some memory management relocated code from the differing id_mm.c
files to a separate place (e.g., near/far memory emulation). It probably
makes more sense to be there anyway. In theory, it'll also be more
convenient to control memory management in certain ways,
if one wishes to.
- This should also let us have primitive (and quite limited)
implementations of functions like malloc/farmalloc and free/farfree.

Note that for the sake of simplicity, EMS (and XMS) is going to be
nuked, but for now it's more-or-less the original code
(although there's no mention of SAVENEARHEAP).

be_cross_mem.c can further be used for storage of resources
like EGAHEAD, although it's not done at the moment.