WARNING - SOMEWHAT BREAKING CATACOMB ABYSS
We're now taking advantage of be_cross_mem.c functions in-game.
Hopefully there are no unexpected glitches, other than what's
documented here.
This includes clean-ups to id_mm.c functions, fixes to be_cross_mem.c
and misc. changes to function references.

KNOWN LIMITATION: Catacomb Abyss is expected to fail to start, with
not enough memory (unless EMULATED_FAR_PARAGRAPHS is increased).
This is because not all shapes in the INTRO are freed.