Hopefully, high audio latency shouldn't lead to a low framerate anymore (reproduced on Android).
With relatively few code changes/additions, and currently controlled by
the newly added BE_ST_FILL_AUDIO_IN_MAIN_THREAD macro, this basically
moves the responsibility of generating the audio samples to the main
thread. It actually uses BE_ST_PrepareForManualAudioCallbackCall, as
already done while the sound subsystem is disabled. Main difference
is that the samples shall be passed to the SDL callback thread now.

This is done with a shared buffer. Its size is calclulated
relatively to our desired SDL callback thread buffer size,
unsign the new "sndInterThreadBufferRatio" setting
(an integer value) as the ratio.

The larger this buffer is, the less chances samples will be
lost (which may lead to audible crackles), although a larger buffer
also requires more memory (well, somewhat). Also, if the SDL callback
isn't exactly in sync with the main thread's SDL_GetTicks based
timing, a few samples may occassionally have to be
skipped (if the SDL thread is too slow), or a small amount
of silence has to be generated (if the thread is too fast).

There are still a few technical issues remaining.