Use a pre-calculated sintable array for the Catacombs, the values of which being present in the newly added fixed_sintable.c source.
Catacomb 3-D v1.22 was run in a debugger-enabled DOSBox build,
and the sintable array was originally dumped from it.

This fixes possible rendering error and standstill projectile bug,
both of which being related to floating point precision differences.
These were reproduced in case sintable[90] was set to GLOBAL1
i.e., 0x10000, rather than 0xFFFF as in the case of the
original EXE. Same could apply to other sintable values.

At least for the projectile itself, this actually exposed
a limitation of FixedByFrac, used in T_Pshot.