Wolf3D: Add stdint.h definitions and additional macros, all being prefixed with id0_, like id0_int_t for the 16-bit signed int type. Adjust definitions of types like byte within id_head.h and id_heads.h, and comment out unused type definitions. This doesn't yet switch to the alternative definitions for all code.
The original plan was to use definitions outside of id's codebase this
time, with a different prefix for standard types, say be0_int_t for int.
However, this could make a bit less sense for types originally defined
in id's own codebase, like boolean or word. It can also be easy to
get confused if two separate prefixes are used for differing types.