be_gamedefs_prep.c restructuring and additions:
While it's still a bit of a mess, there should hopefully be less
code duplication, taking advantage of separate functions.
The great deal of the code in BE_Cross_PrepareGameInstallations
responsible for inspecting specific installation paths moved to
a separate function, BEL_Cross_CheckForKnownInstallations, which
is also using a few helper functions. For more practical changes:
- Try to detect a Windows GOG.com installation of the Catacombs
using a new registry key, in case a later installer was used.
- Detect Wolfenstein 3D and Spear of Destiny from GOG.com. (Thanks WayneCa!)
- Detect Super 3-D Noah's Ark from GOG.com. (Thanks chungy!)