Added the functions BE_ST_EGAUpdateGFXBufferFrom4bitsPixelInPairs, BE_ST_VGAUpdateGFXBufferFrom8bitsPixelInPairs.
They're the same as, e.g., E_ST_EGAUpdateGFXBufferFrom4bitsPixel,
but getting a number of byte pairs to update.
The intention is using these as a drop-in replacement for
a "rep stosw" X86 loop when used for writing to video memory,
especially when the number of byte pairs may be 0x8000.